Remove the need for nbsp's before h1's
Margin can be added to h1's instead of using nbsp's, but the title of
the page is also in an h1 and doesn't need such space.  A hacky solution
is to modify the layout to include the title in an h1 with a new class,
and add a css rule for that class to remove the extra space.  Then the
space can be added to all other h1's.  This means that the title heading
and nbsp's can be removed from the markdown files.

This does slightly break the title of the web page.  Before the page
titles were <page title> --- The Tectonic Project, but now they are just
<page title>.  Fixing this is the next step.